Fix missing initialization of buffer_std field in _bt_newroot().
This would only have any effect if the old root page needed to have
a full-page image made (ie, this was the first mod to it since a
checkpoint), *and* if the byte left uninitialized chanced to contain
zero.  In that case the WAL code would fail to remove the "hole" from
the full-page image, which would bloat the WAL log a bit but not have
any effect worse than that.  Found by buildfarm member skink, whose
valgrind run noticed the use of an uninitialized value.  Apparently
timing in the regression tests is such that the triggering condition
is rare, or valgrind testing would have seen this before.

Oversight in commit 40dae7ec537c5619fc93ad602c62f37be786d161.  This
bug affects only the 9.4 branch, since in later branches refactoring
of the WAL-log-creation APIs fixed it.

Report: <20160521203101.jp5yxquhhkabvo56@alap3.anarazel.de>